Remove WLAN blobs for the time being
Instead of using mac-update or something similar to alter the MAC
addres within the blob, Oppo appears to be using unique blobs with
the correct MAC for a given device in /persist.

For now, use the blobs in /persist.  Also, WCNSS_cfg.dat is in
the device tree on mako, and there isn't anything proprietary
about a text file, so let's move it to the device tree.

Change-Id: I0a792195e7c25fc3f750036422959e251dcee122